+++ src/utils/supabase/middleware.ts
@@ -44,6 +44,7 @@ export async function updateSession(request: NextRequest) {
         !request.nextUrl.pathname.startsWith('/register') &&
         !request.nextUrl.pathname.startsWith('/forgot-password') &&
         !request.nextUrl.pathname.startsWith('/update-password') &&
+        !request.nextUrl.pathname.startsWith('/api/') &&
         request.nextUrl.pathname !== '/'
     ) {
         // no user, potentially respond by redirecting the user to the login page